+++ packages/web-ui/src/ds/molecules/BlocksEditor/types.ts
@@ -1,0 +1,8 @@
+import { type AnyBlock } from '@latitude-data/constants/simpleBlocks'
+
+export interface BlocksEditorProps {
+  blocks: AnyBlock[]
+  onUpdate: (content: string) => void
+  placeholder: string
+  editable?: boolean
+}
